% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{add_powerplant}
\alias{add_powerplant}
\title{Add power plant in power_plants dataframe and in the simulation source files.}
\usage{
add_powerplant(
  capacity_name,
  unit_count,
  area,
  year_simulated,
  study_id,
  power_plants
)
}
\arguments{
\item{capacity_name}{(str) name of the new capacity. If already exists, build a new one with an specific id as suffix.}

\item{unit_count}{(scalar) units to add}

\item{area}{(str) country code}

\item{study_id}{(str) antares simulation id}

\item{power_plants}{dataframe with the power fleet from make_powerplants_from_antares()}

\item{power_plants_actual}{dataframe with the power fleet actual from make_powerplants_from_antares()}

\item{year_actual}{(int) current simulated year}

\item{s}{(int) loop}
}
\value{
power_plants updated
}
\description{
Add power plant in power_plants dataframe and in the simulation source files.
}
