% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{add_production_unit}
\alias{add_production_unit}
\title{Add a new production unit to a power plant and update data structures}
\usage{
add_production_unit(
  study_id,
  area,
  capacity_name,
  new_unitcount,
  old_unitcount,
  power_plants
)
}
\arguments{
\item{study_id}{(str) The ID of the study}

\item{area}{(str) The country code}

\item{capacity_name}{(str) The name of the capacity to which the new unit will be added}

\item{new_unitcount}{(int) The total number of units in the capacity after adding the new unit}

\item{old_unitcount}{(int) The number of units in the old capacity}

\item{power_plants}{(data frame) A data frame containing information about the power plants in the simulation}
}
\value{
A updated version of the \code{power_plants} data frame
}
\description{
This function adds a new production unit to a power plant and updates the corresponding data structures
}
