% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{analysis_revenues}
\alias{analysis_revenues}
\title{Mother function preparing and computing the analysis of the energy only market for every clusters.}
\usage{
analysis_revenues(power_plants, clusters_not_supported, data_antares, area)
}
\arguments{
\item{power_plants}{(dataframe) with costs data for clusters}

\item{clusters_not_supported}{(str vector) clusters interesting for the analysis of revenues and investments}

\item{data_antares}{antaresRead data for the antares run obtained via read_simulation_data()}

\item{area}{(str) country code}
}
\value{
revenues list of revenues for each cluster
}
\description{
Mother function preparing and computing the analysis of the energy only market for every clusters.
}
