% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{analysis_revenues_clusters}
\alias{analysis_revenues_clusters}
\title{Compute the revenues for a cluster c on the energy only market and take into account the heat remuneration}
\usage{
analysis_revenues_clusters(
  c,
  power_plants,
  data_clusters_modulation,
  data_clusters,
  data_clusters_RES,
  data_batteries,
  spot,
  area
)
}
\arguments{
\item{c}{(str) cluster name}

\item{power_plants}{(dataframe) with costs data for clusters}

\item{data_clusters_modulation}{antaresRead table for custers modulation}

\item{data_clusters}{antaresRead data table for thermal clusters}

\item{data_clusters_RES}{antaresRead data table for renewables clusters}

\item{data_batteries}{antaresRead table for batteries}

\item{spot}{spot prices}
}
\value{
annual & hourly revenues and hourly production for the cluster c
}
\description{
Compute the revenues for a cluster c on the energy only market and take into account the heat remuneration
}
