% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_0_analysis_revenues_total.R
\name{analysis_revenues_total}
\alias{analysis_revenues_total}
\title{Mother function to compute the revenues from energy only market and capacity market}
\usage{
analysis_revenues_total(
  year_simulated,
  area,
  power_plants,
  clusters_not_supported,
  data_antares
)
}
\arguments{
\item{year_simulated}{(str) year simulated}

\item{area}{(str) country code}

\item{power_plants}{(dataframe) of powerfleet hypothesis}

\item{clusters_not_supported}{(str vector)clusters names to compute and consider and revenues / investment decisions}

\item{data_antares}{antares read data from an antares run}
}
\value{
list revenues, power_plants, market_info_supply, auction_info
}
\description{
Mother function to compute the revenues from energy only market and capacity market
}
