% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_capacity_shares_chart}
\alias{build_capacity_shares_chart}
\title{Build capacity shares}
\usage{
build_capacity_shares_chart(power_plants)
}
\arguments{
\item{power_plants}{power plants dataframe}
}
\value{
None
}
\description{
Build capacity shares
}
