% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_crm_curves_chart}
\alias{build_crm_curves_chart}
\title{Build CRM curves (demand & supply)}
\usage{
build_crm_curves_chart(
  graph_dir,
  crm_supply,
  crm_demand,
  auction_info,
  ro_strike,
  zone
)
}
\arguments{
\item{graph_dir}{(str) graphic directory}

\item{crm_demand}{(float) demand in MW}

\item{auction_info}{(list) information of the auction}

\item{ro_strike}{ro strike dataframe}

\item{zone}{zone}

\item{crm_supply_curve}{(data.frame) supply_curve (bid, cumulated certified capacity)}

\item{year_simulated}{(int) year simulation}

\item{s}{s}
}
\value{
None
}
\description{
Build CRM curves (demand & supply)
}
