% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_build_dataset.R
\name{build_dataset}
\alias{build_dataset}
\title{Mother function building all data from antares and hypothesis}
\usage{
build_dataset(
  year_simulated,
  power_plants,
  antares_output_name,
  area,
  study_id
)
}
\arguments{
\item{year_simulated}{(str) year simulated to get antares outputs}

\item{antares_output_name}{(str) name of the output antares corresponding to a run.}

\item{area}{(str) country code}

\item{study_id}{(str) simulation name or id API}
}
\value{
margins and data_antares and clusters filters
}
\description{
Mother function building all data from antares and hypothesis
}
