% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{build_dataset_wrapper}
\alias{build_dataset_wrapper}
\title{Wrapper for buid_dataset function}
\usage{
build_dataset_wrapper(power_plants, study_list, years_to_simulate, outputs)
}
\arguments{
\item{power_plants}{(list) power_plants dataframe hypothesis for each country}

\item{study_list}{(list)}

\item{years_to_simulate}{(int vector)}

\item{outputs}{(list) antares outputs}
}
\value{
dataset (list) antaresRead data and other processed data after the antares run
}
\description{
Wrapper for buid_dataset function
}
