% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_CRM_utils.r
\name{build_demand_curve}
\alias{build_demand_curve}
\title{Simulate the supply curve}
\usage{
build_demand_curve(load = NA, area)
}
\arguments{
\item{load}{(float) load to define the inelastic demand}

\item{area}{(str) country code}
}
\description{
Simulate the supply curve
}
