% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{build_folders_architecture}
\alias{build_folders_architecture}
\title{Creating directories for the study and the user}
\usage{
build_folders_architecture()
}
\value{
None
}
\description{
Creating directories for the study and the user
}
