% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_1_make_outputs.R
\name{build_from_data_outputs}
\alias{build_from_data_outputs}
\title{Reformat data from data_outputs in order to seperate interesting variables}
\usage{
build_from_data_outputs(data_outputs, decisions, max_loop)
}
\arguments{
\item{data_outputs}{outputs recorded from the model}

\item{decisions}{investments/decomissioning decisions}

\item{max_loop}{number maximum of really run loops over years}
}
\value{
list of lists with interesting variables
}
\description{
Reformat data from data_outputs in order to seperate interesting variables
}
