% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_lold_chart}
\alias{build_lold_chart}
\title{Make the graph of the Loss of Load duration evolution}
\usage{
build_lold_chart(LOLD, max_loop)
}
\arguments{
\item{LOLD}{loss of load duration dataframe}

\item{max_loop}{number maximum of really run loops over years}
}
\value{
None
}
\description{
Make the graph of the Loss of Load duration evolution
}
