% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_0_analysis_revenues_total.R
\name{build_market_info_supply}
\alias{build_market_info_supply}
\title{Build the market info supply dataframe}
\usage{
build_market_info_supply(clusters_not_supported, power_plants, revenues, area)
}
\arguments{
\item{clusters_not_supported}{(str vector) clusters not supported}

\item{power_plants}{(dataframe) of clusters costs hypothesis}

\item{revenues}{list of dataframe revenues by cluster}
}
\value{
market info on the supply side
}
\description{
Build the market info supply dataframe
}
