% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_build_parameters.R
\name{build_parameters_input}
\alias{build_parameters_input}
\title{Read yaml config parameters file and set to global environnement}
\usage{
build_parameters_input(update_yaml = FALSE)
}
\value{
None
}
\description{
Read yaml config parameters file and set to global environnement
}
