% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_build_parameters.R
\name{build_parameters_output}
\alias{build_parameters_output}
\title{Build the parameters yaml file in order to record the parameters used}
\usage{
build_parameters_output(toc, start_time)
}
\arguments{
\item{toc}{elapsed time of the study}
}
\value{
write the parameters yaml file in results/study_name folder
}
\description{
Build the parameters yaml file in order to record the parameters used
}
