% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_power_plants_moves_chart}
\alias{build_power_plants_moves_chart}
\title{Build power_plants moves graphs}
\usage{
build_power_plants_moves_chart(power_plants_moves)
}
\arguments{
\item{power_plants_moves}{List of investments & decommissioning decisions}
}
\value{
power_plants_moves reshaped
}
\description{
Build power_plants moves graphs
}
