% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_powerfleet_chart}
\alias{build_powerfleet_chart}
\title{Make the graph of the powerfleet evolution by cluster}
\usage{
build_powerfleet_chart(graph_dir, power_plants, zone)
}
\arguments{
\item{graph_dir}{graphic directory}

\item{power_plants}{power fleet list of dataframe}

\item{zone}{country code}
}
\value{
None
}
\description{
Make the graph of the powerfleet evolution by cluster
}
