% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_powerfleet_europe_chart}
\alias{build_powerfleet_europe_chart}
\title{Build power plants graphs aggregated at european scale / country scale}
\usage{
build_powerfleet_europe_chart(power_plants, max_loop)
}
\arguments{
\item{power_plants}{power_plants dataframe}

\item{max_loop}{number maximum of really run loops over years}
}
\value{
None
}
\description{
Build power plants graphs aggregated at european scale / country scale
}
