% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_production_chart}
\alias{build_production_chart}
\title{Build production graphs}
\usage{
build_production_chart(graph_dir, production, zone)
}
\arguments{
\item{graph_dir}{graphic directory}

\item{production}{production dataframe}

\item{zone}{country code}
}
\value{
None
}
\description{
Build production graphs
}
