% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_revenues_chart}
\alias{build_revenues_chart}
\title{Make the graph of revenues evolution by cluster}
\usage{
build_revenues_chart(graph_dir, revenues, zone)
}
\arguments{
\item{graph_dir}{graphic directory}

\item{revenues}{revenues list of dataframe}

\item{zone}{country code}
}
\value{
None
}
\description{
Make the graph of revenues evolution by cluster
}
