% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_make_graphs.R
\name{build_spot_chart}
\alias{build_spot_chart}
\title{Make the graph of the spot evolution}
\usage{
build_spot_chart(spot, max_loop)
}
\arguments{
\item{spot}{dataframe of spot prices}

\item{max_loop}{number maximum of really run loops over years}
}
\value{
None
}
\description{
Make the graph of the spot evolution
}
