% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_CRM_utils.r
\name{build_stairs_curve}
\alias{build_stairs_curve}
\title{Simulate the supply curve}
\usage{
build_stairs_curve(market_info_supply, total_capacity)
}
\arguments{
\item{market_info_supply}{(data.table) dataframe from compute_bids()}

\item{total_capacity}{(float) the total certified capacity}
}
\description{
Simulate the supply curve
}
