% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_build_dataset.R
\name{build_study_list}
\alias{build_study_list}
\title{Build the study list for local and API}
\usage{
build_study_list(params_study_list)
}
\arguments{
\item{params_study_list}{}
}
\value{
study_list (list) study ids list
}
\description{
Build the study list for local and API
}
