% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_build_parameters.R
\name{build_user_parameter}
\alias{build_user_parameter}
\title{Build user_params file in yaml format}
\usage{
build_user_parameter()
}
\value{
None. File is written.
}
\description{
Build user_params file in yaml format
}
