% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_data_investment.R
\name{build_vector_npv}
\alias{build_vector_npv}
\title{Build a vector of size year_start : time_horizon with 0 for years not simulated}
\usage{
build_vector_npv(revenues_vector, years_to_simulate, year_start, replace = T)
}
\arguments{
\item{revenues_vector}{(scalar vector) revenues vector}

\item{years_to_simulate}{(int vector) years to simulate}

\item{year_start}{(int) starting year of the NPV vector}

\item{replace}{(bool) if TRUE replace by zero when year is not in years_to_simulate}
}
\value{
revenues vector prepared for NPV computation
}
\description{
Build a vector of size year_start : time_horizon with 0 for years not simulated
}
