% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{check_antares_dataset}
\alias{check_antares_dataset}
\title{Make the first check, to assert simulation dataset is suitable for AntaresInvest}
\usage{
check_antares_dataset(
  study_id,
  power_plants,
  power_plants_actual,
  area,
  lowering = TRUE,
  editing_outputs = TRUE,
  filling_info_cluster = TRUE
)
}
\arguments{
\item{study_id}{(str) antares simulation id}

\item{power_plants}{(dataframe) powerplants actual dataframe with all the hypothesis}

\item{area}{(str) country code}
}
\value{
None
}
\description{
Make the first check, to assert simulation dataset is suitable for AntaresInvest
}
