% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{check_antares_output}
\alias{check_antares_output}
\title{Check if antares_output_name is coherent with launch time of antares run}
\usage{
check_antares_output(study_id, antares_output_name)
}
\arguments{
\item{study_id}{(str)}

\item{antares_output_name}{(str)}
}
\value{
None
}
\description{
Check if antares_output_name is coherent with launch time of antares run
}
