% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_availability.R
\name{compute_availability_enr}
\alias{compute_availability_enr}
\title{Compute certified capacity for renewables energy sources}
\usage{
compute_availability_enr(area, load, clusters_enr, availability_renewables)
}
\arguments{
\item{area}{(str) country code}

\item{load}{demand (hourly level)}

\item{clusters_enr}{(str vector) RES clusters names candidates to revenues analysis and investments}

\item{study_id}{(str) id of the antares simulation}
}
\value{
certified capacity for ENR (wind and solar) on PP1 days
}
\description{
Compute certified capacity for renewables energy sources
}
