% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_availability.R
\name{compute_availability_interconnections}
\alias{compute_availability_interconnections}
\title{Compute the availability of the interconnections}
\usage{
compute_availability_interconnections(area, links, LOLD)
}
\arguments{
\item{area}{(str) country code}

\item{links}{(antares datatable) links flow for neighbours}

\item{LOLD}{(antares datatable) Loss of load duration}
}
\value{
capacity_certified for every interconections of a country
}
\description{
Compute the availability of the interconnections
}
