% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_availability.R
\name{compute_availability_thermal}
\alias{compute_availability_thermal}
\title{Compute availability on PP1 days of not supported clusters}
\usage{
compute_availability_thermal(
  load,
  clusters_thermal,
  power_plants,
  availability_thermal
)
}
\arguments{
\item{load}{(scalar) demand (hourly level)}

\item{clusters_thermal}{(str vector) thermal clusters names candidates to revenues analysis and investments}

\item{power_plants}{(dataframe) with costs data for clusters}

\item{availability_thermal}{(data.table) availability data for thermal clusters from readAntares}
}
\value{
certified capacity by cluster
}
\description{
Compute availability on PP1 days of not supported clusters
}
