% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_availability.R
\name{compute_availability_total}
\alias{compute_availability_total}
\title{Compute availability on PP1 days}
\usage{
compute_availability_total(power_plants, clusters_not_supported, data_antares)
}
\arguments{
\item{power_plants}{dataframe with costs data for clusters}

\item{clusters_not_supported}{(str vector) clusters names candidates to revenues analysis and investments}

\item{data_antares}{(list of antare data table) all useful extracted data from the antares run}

\item{study_id}{(str) antares simulation id}
}
\value{
Thermal availability for every clusters (ie capacity certifiee)
}
\description{
Compute availability on PP1 days
}
