% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_CRM.R
\name{compute_bids}
\alias{compute_bids}
\title{Compute the bids of the supply side and make the market_info_supply side dataframe}
\usage{
compute_bids(
  clusters_not_supported,
  power_plants,
  capacity_certified,
  market_info_supply,
  margin,
  obligation,
  revenues,
  spot,
  zone
)
}
\arguments{
\item{clusters_not_supported}{(str vector) clusters names candidates to the revenues analysis and investments.}

\item{power_plants}{(dataframe) of costs hypothesis}

\item{capacity_certified}{(scalar) certified capacity by clusters}

\item{market_info_supply}{(datatable) summary of information on supply side}

\item{revenues}{(list of dataframes)}

\item{spot}{}

\item{zone}{(str) zone}
}
\value{
market_info_supply a dataframe with info concerning the supply side (revenues, bids capacity.. )
}
\description{
Compute the bids of the supply side and make the market_info_supply side dataframe
}
