% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{compute_dsr_operation_cost}
\alias{compute_dsr_operation_cost}
\title{Compute the om costs of dsrs}
\usage{
compute_dsr_operation_cost(capacity, cost_dsr)
}
\arguments{
\item{capacity}{capacity}

\item{cost_dsr}{table of dsrs costs.}
}
\value{
om cost of dsrs
}
\description{
Compute the om costs of dsrs
}
