% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_price_cap.r
\name{compute_max_hourly_spot}
\alias{compute_max_hourly_spot}
\title{Compute max hourly spot prices in Europe}
\usage{
compute_max_hourly_spot(antares_output_name, study_id)
}
\arguments{
\item{antares_output_name}{(str)  name of the output antares corresponding to a run.}

\item{study_id}{(str) the name of the antares simulation}
}
\value{
Hourly max spot prices in Europe
}
\description{
Compute max hourly spot prices in Europe
}
