% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_CRM_utils.r
\name{compute_obligation}
\alias{compute_obligation}
\title{Compute the obligation for capacity market}
\usage{
compute_obligation(data_antares, zone)
}
\arguments{
\item{data_antares}{list of antaresRead datatable with all data from output antares simulation}

\item{zone}{(str) country code}
}
\value{
obligation
}
\description{
Compute the obligation for capacity market
}
