% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{create_cluster_from_existing}
\alias{create_cluster_from_existing}
\title{Add a new power plant to a simulation and update data structures.}
\usage{
create_cluster_from_existing(
  area,
  study_id,
  capacity_name,
  new_unitcount,
  power_plants
)
}
\arguments{
\item{area}{(str) The country code}

\item{study_id}{(str) The ID of the study}

\item{capacity_name}{(str) The name of the new capacity If the capacity already exists, a new one will be created with a specific ID as a suffix}

\item{new_unitcount}{(int) The number of units in the new capacity}

\item{power_plants}{(data frame) A data frame containing information about the power plants in the simulation}
}
\value{
A updated version of the \code{power_plants} data frame.
}
\description{
Add a new power plant to a simulation and update data structures.
}
