% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{decorate_message}
\alias{decorate_message}
\title{Print decorated message}
\usage{
decorate_message(message)
}
\arguments{
\item{message}{the message to decorate}
}
\description{
Print decorated message
}
