% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{delete_all_variants_of_a_study}
\alias{delete_all_variants_of_a_study}
\title{Detele all variants of a given study}
\usage{
delete_all_variants_of_a_study(study_id, year)
}
\arguments{
\item{study_id}{study_id of the original study}

\item{year}{year of the simulation (for verbose)}
}
\value{
None
}
\description{
Detele all variants of a given study
}
