% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{delete_outputs_antares}
\alias{delete_outputs_antares}
\title{Remove outputs of antares runs when finished to recover spaces on disk}
\usage{
delete_outputs_antares(force = FALSE)
}
\value{
None
}
\description{
Remove outputs of antares runs when finished to recover spaces on disk
}
