% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{delete_study_antares_web}
\alias{delete_study_antares_web}
\title{Delete an antares web study thanks to antares_web API}
\usage{
delete_study_antares_web(study_id, variant_name, year)
}
\arguments{
\item{study_id}{study_id of the simulation we are launching}

\item{variant_name}{name of the variant we will delete}

\item{year}{year of simulation}
}
\value{
None
}
\description{
Delete an antares web study thanks to antares_web API
}
