% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_price_cap.r
\name{edit_price_cap}
\alias{edit_price_cap}
\title{Edit price cap in the simulation}
\usage{
edit_price_cap(
  price_cap = NA,
  study_id,
  restore_mode = FALSE,
  reading_mode = FALSE
)
}
\arguments{
\item{price_cap}{(scalar) new price cap for energy only market}

\item{study_id}{(str) simulation name}

\item{restore_mode}{(bool) if yes or no, restore the initial price cap}
}
\value{
None or the initial price cap if reading mode = TRUE
}
\description{
Edit price cap in the simulation
}
