% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{edit_ts_input_cluster}
\alias{edit_ts_input_cluster}
\title{Edit time series for thermal cluster}
\usage{
edit_ts_input_cluster(study_id, area, capacity_name, value)
}
\arguments{
\item{study_id}{study_id of study you want to edit}

\item{area}{area containing link or cluster}

\item{value}{value to use with operation; e.g. if operation "=" and value "10" will set all time series value equal to 10}

\item{cluster_name}{name of cluster (or link)}

\item{operation}{which operation to apply to time series => '=' ; '*' ; '+'}
}
\value{
None
}
\description{
Edit time series for thermal cluster
}
