% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{energy_revenues_cluster}
\alias{energy_revenues_cluster}
\title{Compute energy only revenues on hourly level by cluster}
\usage{
energy_revenues_cluster(c, marginal_cost, production, spot)
}
\arguments{
\item{c}{(str) cluster name}

\item{marginal_cost}{(scalar) variable cost (ie the bid on the energy only market)}

\item{production}{production (hourly level)}

\item{spot}{spot prices (hourly level)}
}
\value{
energy only revenues (hourly level)
}
\description{
Compute energy only revenues on hourly level by cluster
}
