% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{ENP}
\alias{ENP}
\title{Expected net profit}
\usage{
ENP(year_end, year_start, revenues_energy, revenues_crm, operational_cost)
}
\arguments{
\item{year_end}{the ending year of the expected net profit computation (different from time horizon)}

\item{year_start}{the starting year of the computation}

\item{revenues_energy}{energy only revenues (euros/MW installed)}

\item{revenues_crm}{capacity market revenues (euros/MW installed)}

\item{operational_cost}{operational costs (euros/MW installed)}
}
\value{
expected net profit of the cluster
}
\description{
Expected net profit
}
