% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{extract_ts_clusters}
\alias{extract_ts_clusters}
\title{Extract time series of a cluster}
\usage{
extract_ts_clusters(capacity_name, study_id, coefficient)
}
\arguments{
\item{capacity_name}{(str) name of the new capacity. If already exists, build a new one with an specific id as suffix.}

\item{study_id}{(str) id of study}

\item{coefficient}{(float) coefficient to extract and scale the timeseries}
}
\value{
timeseries (list) of every timeseries of the cluster
}
\description{
Extract time series of a cluster
}
