% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{generate_new_capacity_name}
\alias{generate_new_capacity_name}
\title{Generate the new capacity name when adding a cluster with the same property than an existing one}
\usage{
generate_new_capacity_name(capacity_name, power_plants)
}
\arguments{
\item{capacity_name}{(str) initial capacity name}

\item{power_plants}{(dataframe) with the power fleetfrom make_powerplants_from_antares()}
}
\value{
new_capacity_name
}
\description{
Generate the new capacity name when adding a cluster with the same property than an existing one
}
