% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{get_all_variants_of_study}
\alias{get_all_variants_of_study}
\title{Get all the variants with their level (essential to delete study), recursively}
\usage{
get_all_variants_of_study(l, v_list = list(), level = 0)
}
\arguments{
\item{l}{list}

\item{v_list}{variant list}

\item{level}{level of the variant (level=0 is the reference study)}
}
\value{
None
}
\description{
Get all the variants with their level (essential to delete study), recursively
}
