% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{get_cluster_type}
\alias{get_cluster_type}
\title{Get the cluster type of any antares cluster based on its name. Mandatory to determine this cluster cost automatically}
\usage{
get_cluster_type(cluster)
}
\arguments{
\item{cluster}{(str) name of cluster e.g fr_oil_chp}
}
\value{
cluster_type (str): the type of cluster
}
\description{
Get the cluster type of any antares cluster based on its name. Mandatory to determine this cluster cost automatically
}
