% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{get_clusters_filters}
\alias{get_clusters_filters}
\title{Build the clusters interesting list}
\usage{
get_clusters_filters(zone, cluster_list)
}
\arguments{
\item{zone}{(str) country code}

\item{cluster_list}{(str vector) dataframe with the cluster of the zone and their name}
}
\value{
clusters_filters for one year simulated (the one of data_antares)
}
\description{
Build the clusters interesting list
}
