% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{get_costs_clusters}
\alias{get_costs_clusters}
\title{Build the costs clusters dataframe}
\usage{
get_costs_clusters(zone, power_plants)
}
\arguments{
\item{zone}{(str) country code}

\item{power_plants}{(df) list dataframe of the hypothesis of the powerfleet}

\item{study_id}{(str) id of the simulation antares}
}
\value{
updated powerplants hypothesis dataframe
}
\description{
Build the costs clusters dataframe
}
