% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_data_investment.R
\name{get_data_for_metric}
\alias{get_data_for_metric}
\title{Get data processed for metric (NPV, PI, IRR) computation}
\usage{
get_data_for_metric(
  cluster,
  years_to_simulate,
  power_plants,
  dataset_revenues,
  area
)
}
\arguments{
\item{cluster}{(str) cluster name}

\item{years_to_simulate}{(vector of int) with years to simulate}

\item{power_plants}{(dataframe) with costs data}

\item{dataset_revenues}{(list of dataframe) with the revenues of each cluster for each year in years_to_simulate}

\item{area}{(str) country code}
}
\value{
list of data : revenues eom, revenues crm, costs om, costs investissements,
}
\description{
Get data processed for metric (NPV, PI, IRR) computation
}
