% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_1_make_outputs.R
\name{get_data_outputs_model}
\alias{get_data_outputs_model}
\title{Get data outputs to store/record data from the model}
\usage{
get_data_outputs_model(
  dataset = NA,
  dataset_revenues = NA,
  power_plants = NA,
  year_actual = NA,
  data_outputs = NA,
  continue = NA,
  investment_metrics = NA,
  s = NA
)
}
\arguments{
\item{dataset}{dataset from antares runs outputs}

\item{dataset_revenues}{(list of dataframe) revenues by cluster and year returned by analysis_revenues_total}

\item{power_plants}{power_plants dataframe}

\item{year_actual}{(int) actual year of the model}

\item{data_outputs}{(list of dataframe) where results will be stored}

\item{investment_metrics}{(dataframe) investment metrics for the year actual and loop s}

\item{s}{(int) loop number on the actual year}

\item{setup}{(bool) if TRUE, function returns empty dataframe ready to store data}
}
\value{
variables to store data from the model : power_plants_moves & data_outputs
}
\description{
Get data outputs to store/record data from the model
}
