% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{get_data_outputs_wrapper}
\alias{get_data_outputs_wrapper}
\title{Wrapper for get_data_outputs}
\usage{
get_data_outputs_wrapper(
  dataset,
  dataset_revenues,
  power_plants,
  year_actual,
  data_outputs,
  s,
  investment_metrics,
  continue
)
}
\arguments{
\item{dataset}{(list) antaresRead data and other processed data after the antares run}

\item{dataset_revenues}{(list) computed revenues for each cluster for each country}

\item{power_plants}{(list) power_plants dataframe hypothesis list for each country}

\item{year_actual}{(int) actual year}

\item{data_outputs}{(list) outputs recovered from the model}

\item{s}{(int) loop}

\item{investment_metrics}{(list) investments metrics dataframe for each country}

\item{continue}{(int) continue variable}
}
\value{
data_outputs (list) outputs recovered from the model updated
}
\description{
Wrapper for get_data_outputs
}
