% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_link.R
\name{get_neighbours_links}
\alias{get_neighbours_links}
\title{Get the links of the neighbours area of the area considered}
\usage{
get_neighbours_links(area, study_id)
}
\arguments{
\item{area}{(str) country code}

\item{study_id}{(str) study id of the simulation antares (if API mode)}
}
\value{
links_neighbours (str vector)
}
\description{
Get the links of the neighbours area of the area considered
}
