% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{heat_revenues_cluster}
\alias{heat_revenues_cluster}
\title{Compute heat revenues}
\usage{
heat_revenues_cluster(c, heat_production, power_plants)
}
\arguments{
\item{c}{(str) cluster name}

\item{heat_production}{heat production (hourly level)}

\item{power_plants}{power_plants dataframe hypothesis}
}
\value{
heat revenues
}
\description{
Compute heat revenues
}
