% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{IRR}
\alias{IRR}
\title{Internal rate of return}
\usage{
IRR(
  revenues_energy,
  revenues_crm,
  investment_cost,
  operational_cost,
  year_start,
  years_to_simulate
)
}
\arguments{
\item{revenues_energy}{energy only revenues (euros/MW installed)}

\item{revenues_crm}{capacity market revenues (euros/MW installed)}

\item{investment_cost}{investment costs (euros/MW installed)}

\item{operational_cost}{operational costs (euros/MW installed)}

\item{current_year}{the starting year of the computation}
}
\value{
Internal rate of return
}
\description{
Internal rate of return
}
