% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_1_make_outputs.R
\name{make_outputs}
\alias{make_outputs}
\title{Mother function of outputs (ie graphs, and outputs files)}
\usage{
make_outputs(data_outputs, area, decisions, max_loop)
}
\arguments{
\item{data_outputs}{(list)outputs from the model}

\item{area}{(str) country code}

\item{decisions}{(list) investment decisions}

\item{max_loop}{(int) number maximum of really run loops over years}
}
\value{
Make automated graphics and charts in the results/study_name/graphiques folder
}
\description{
Mother function of outputs (ie graphs, and outputs files)
}
