% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{make_outputs_wrapper}
\alias{make_outputs_wrapper}
\title{Wrapper of multiple make_outputs functions}
\usage{
make_outputs_wrapper(data_outputs, decisions, max_loop)
}
\arguments{
\item{data_outputs}{(list) Recovered outputs of the model by get_data_outputs()}

\item{decisions}{(list) investment decisions}

\item{max_loop}{(int) max loop reached}
}
\value{
None
}
\description{
Wrapper of multiple make_outputs functions
}
