% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{make_powerplants_from_antares}
\alias{make_powerplants_from_antares}
\title{Make power_plants dataframe based on thermal clusters list ini file and returns it.}
\usage{
make_powerplants_from_antares(study_id, year_simulated, zone)
}
\arguments{
\item{study_id}{(str) id of the year simulated}

\item{year_simulated}{(str) year simulated (for verbose)}

\item{area}{(str) country code}
}
\value{
power_plants dataframe
}
\description{
Make power_plants dataframe based on thermal clusters list ini file and returns it.
}
