% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_price_cap.r
\name{make_price_caps_moves}
\alias{make_price_caps_moves}
\title{Compute price cap following ACER decision of 14 November 2017}
\usage{
make_price_caps_moves(antares_output_name, year_actual)
}
\arguments{
\item{antares_output_name}{(str) output of the antares run}

\item{year_actual}{(str) name of the antares simulation}

\item{initial_price_cap}{(scalar) initial price cap}
}
\value{
price cap evolution
}
\description{
Compute price cap following ACER decision of 14 November 2017
}
