% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{make_years}
\alias{make_years}
\title{Make years wrapper}
\usage{
make_years(antares_actual)
}
\arguments{
\item{antares_actual}{(str) the year actual}
}
\value{
(list) with year actual and the years to be simulated
}
\description{
Make years wrapper
}
