% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{modulate_capacity}
\alias{modulate_capacity}
\title{Update the capacity at each time knowing capacity modulation factor}
\usage{
modulate_capacity(
  power_plants,
  clusters_not_supported,
  data_clusters_modulation
)
}
\arguments{
\item{power_plants}{(datafrmame) with costs data for clusters}

\item{clusters_not_supported}{(str vector)}

\item{data_clusters_modulation}{antaresRead data.table for custers modulation}
}
\value{
data_clusters_modulation updated with capacity modulated
}
\description{
Update the capacity at each time knowing capacity modulation factor
}
