% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{NPV}
\alias{NPV}
\title{Compute NPV (unit /MW installed -> all costs and revenues are per MW installed)}
\usage{
NPV(
  revenues_energy,
  revenues_crm,
  investment_cost,
  operational_cost,
  year_start,
  years_to_simulate,
  discount_rate
)
}
\arguments{
\item{revenues_energy}{energy only revenues (euros/MW installed)}

\item{revenues_crm}{capacity market revenues (euros/MW installed)}

\item{investment_cost}{investment costs (euros/MW.year installed)}

\item{operational_cost}{operational costs (euros/MW.year installed)}

\item{year_start}{the starting year of the computation}

\item{discount_rate}{discount rate}
}
\value{
NPV
}
\description{
Compute NPV (unit /MW installed -> all costs and revenues are per MW installed)
}
