% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{parallel_run_antares_web}
\alias{parallel_run_antares_web}
\title{Launch in parralel several simulation on antares Web}
\usage{
parallel_run_antares_web(years_to_simulate, sub_mode, mc_list, run_name)
}
\arguments{
\item{years_to_simulate}{years 202K to simulate}

\item{sub_mode}{sub_mode ("fast")}

\item{mc_list}{monte carlo years list}

\item{run_name}{name of study}
}
\value{
outputs
}
\description{
Launch in parralel several simulation on antares Web
}
