% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{parallelisation_antares}
\alias{parallelisation_antares}
\title{Parallelisation of antares run}
\usage{
parallelisation_antares(
  nb_cores,
  solver_path,
  years_to_simulate,
  antares_simu_dir,
  sub_mode,
  mc_list,
  run_name,
  antares_mode,
  antares_run_save
)
}
\arguments{
\item{nb_cores}{cores to use for the calculation}

\item{solver_path}{path of antares solver}

\item{years_to_simulate}{years to simulate}

\item{antares_simu_dir}{antares simulation directory}

\item{sub_mode}{fast or accuratev lancer 'apAPImatin matib}

\item{mc_list}{montecarlo list}

\item{run_name}{name of the run_name without the default YYMMDD-HHmm}

\item{antares_run_save}{option to make a run antares not allowed for deleting automatically}
}
\value{
output names of the runs.

None
}
\description{
Parallelisation of antares run
}
