% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_build_dataset.R
\name{parallelisation_build_dataset}
\alias{parallelisation_build_dataset}
\title{Parallelized load of the dataset for a list of years_to_simulate and areas}
\usage{
parallelisation_build_dataset(years_to_simulate, areas)
}
\arguments{
\item{years_to_simulate}{(int vector) Years to simulate}

\item{areas}{(str vector) Areas}
}
\value{
dataset
}
\description{
Parallelized load of the dataset for a list of years_to_simulate and areas
}
