% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_3_make_outputs_utils.r
\name{parallelisation_make_outputs}
\alias{parallelisation_make_outputs}
\title{Parallelized outputs creation (graphs & Excel sheets)}
\usage{
parallelisation_make_outputs(areas)
}
\arguments{
\item{areas}{Areas}
}
\value{
outputs of the model
}
\description{
Parallelized outputs creation (graphs & Excel sheets)
}
