% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{PI}
\alias{PI}
\title{Profitability index define as NPV / IC}
\usage{
PI(NPV, investment_cost)
}
\arguments{
\item{NPV}{Net present value}

\item{investment_cost}{investment costs (euros/MW installed)}
}
\value{
Profitability index
}
\description{
Profitability index define as NPV / IC
}
