% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{prepare_investment_metrics}
\alias{prepare_investment_metrics}
\title{Prepare investment metrics for the investment model}
\usage{
prepare_investment_metrics(
  year_actual,
  years_to_simulate,
  clusters_not_supported,
  dataset_revenues,
  dataset,
  power_plants,
  area
)
}
\arguments{
\item{year_actual}{the actual year of the model}

\item{years_to_simulate}{(int vector) years to simulate}

\item{clusters_not_supported}{(str vector) interesting clusters for revenues analysis and investments decisions}

\item{dataset_revenues}{(list of datatable) revenues and production list by clusters and years}

\item{dataset}{(list of datatable) dataset of outputs runs from antares from build dataset}

\item{power_plants}{(dataframe) powerfleet hypothesis dataframe}

\item{area}{(str) country code}
}
\value{
metrics table, dataframe of revenues, candidates decommissioning
}
\description{
Prepare investment metrics for the investment model
}
