% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_read_antares_data.R
\name{read_simulation_data}
\alias{read_simulation_data}
\title{Read the simulation data from the antares_output_name run}
\usage{
read_simulation_data(study_id, antares_output_name, zone)
}
\arguments{
\item{study_id}{(str) the name of the antares simulation}

\item{antares_output_name}{(str) name of the output antares corresponding to a run.}

\item{zone}{(str) country code}
}
\value{
data_antares
}
\description{
Read the simulation data from the antares_output_name run
}
