% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_3_make_outputs_utils.r
\name{record_investment_metrics}
\alias{record_investment_metrics}
\title{Function to record during the model runs the investment metrics as excel files.}
\usage{
record_investment_metrics()
}
\description{
Function to record during the model runs the investment metrics as excel files.
}
