% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_data_investment.R
\name{reformat_df_revenues_total}
\alias{reformat_df_revenues_total}
\title{Build the dataframe of total net revenues by cluster x years}
\usage{
reformat_df_revenues_total(revenues_total_net, years_to_simulate)
}
\arguments{
\item{revenues_total_net}{(dataframe) of net revenues}

\item{years_to_simulate}{(vector of int) with years to simulate}
}
\value{
dataframe of net revenues by clusters and years prepared for investment decision
}
\description{
Build the dataframe of total net revenues by cluster x years
}
