% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{remove_powerplant}
\alias{remove_powerplant}
\title{Remove power plant mother function from power_plants dataframe and from the simulation source files.}
\usage{
remove_powerplant(
  capacity_name,
  unit_count,
  power_plants,
  area,
  year_simulated,
  study_id
)
}
\arguments{
\item{capacity_name}{(str) name of the capacity to remove}

\item{unit_count}{(int) units to remove}

\item{power_plants}{(dataframe) with the power fleetfrom make_powerplants_from_antares()}

\item{area}{(str) country code}

\item{year_simulated}{(int) year of simulation to update}

\item{study_id}{(str) antares simulation id}
}
\value{
power_plants dataframe updated
}
\description{
Remove power plant mother function from power_plants dataframe and from the simulation source files.
}
