% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{remove_production_unit}
\alias{remove_production_unit}
\title{Decomission n units of a capacity}
\usage{
remove_production_unit(
  area,
  unitcount,
  new_unitcount,
  capacity_name,
  study_id,
  year_simulated
)
}
\arguments{
\item{area}{(str) country code}

\item{unitcount}{(int) unitcount decomissioned}

\item{new_unitcount}{(int) new unit count before retirement}

\item{capacity_name}{(str) name of the capacity}

\item{study_id}{(str) antares simulation id}

\item{year_simulated}{(str) year simulated}
}
\value{
powerplants dataframe updated
}
\description{
Decomission n units of a capacity
}
