% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_3_make_outputs_utils.r
\name{remove_useless_loops}
\alias{remove_useless_loops}
\title{Remove loops stored uselessly between max_loop and iteration_max}
\usage{
remove_useless_loops(max_loop, data_outputs)
}
\arguments{
\item{max_loop}{number maximum of really run loops over years}
}
\description{
Remove loops stored uselessly between max_loop and iteration_max
}
