% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{rename_dataset}
\alias{rename_dataset}
\title{Rename the second level of dataset with areas names}
\usage{
rename_dataset(dataset, years_to_simulate, areas)
}
\arguments{
\item{dataset}{the 2 leveled dataset list to rename with first level : years, second level : countries}

\item{years_to_simulate}{(str vector) the years to simulate}

\item{areas}{(str vector)}
}
\value{
dataset updated with areas name
}
\description{
Rename the second level of dataset with areas names
}
