% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{restore_antares}
\alias{restore_antares}
\title{Wrapper of the restore antares function either or not for antares Web}
\usage{
restore_antares(study_list)
}
\value{
None
}
\description{
Wrapper of the restore antares function either or not for antares Web
}
