% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{restore_antares_simulation}
\alias{restore_antares_simulation}
\title{Save and restore the antares power fleet.}
\usage{
restore_antares_simulation(
  save_id = NA,
  saving_mode = FALSE,
  end_etude = FALSE
)
}
\arguments{
\item{save_id}{number of the save}

\item{saving_mode}{mode to save or restore}

\item{end_etude}{if true, every saves are erased and the power fleet are all restored.}
}
\value{
None
}
\description{
Save and restore the antares power fleet.
}
