% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{restore_antares_simulation_save_mode}
\alias{restore_antares_simulation_save_mode}
\title{Save the antares simulation files used by the model}
\usage{
restore_antares_simulation_save_mode(area, save_id)
}
\arguments{
\item{area}{country code}

\item{save_id}{id of the save}
}
\value{
None
}
\description{
Save the antares simulation files used by the model
}
