% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{restore_antares_web_simulation}
\alias{restore_antares_web_simulation}
\title{Remove all the variants of the studies stored in study_list}
\usage{
restore_antares_web_simulation()
}
\value{
None
}
\description{
Remove all the variants of the studies stored in study_list
}
