% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{retirement_capacity}
\alias{retirement_capacity}
\title{Retire completely a capacity}
\usage{
retirement_capacity(
  capacity_name,
  area,
  old_unitcount,
  study_id,
  year_simulated
)
}
\arguments{
\item{capacity_name}{(str) name of the capacity}

\item{area}{(str) country code}

\item{old_unitcount}{(int) old unit count before retirement}

\item{study_id}{(str) antares simulation id}

\item{year_simulated}{(str) year simulated}
}
\value{
powerplants dataframe updated
}
\description{
Retire completely a capacity
}
