% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{run_antares}
\alias{run_antares}
\title{Wrapper of the run antares simulation : local and antares_web}
\usage{
run_antares(years_to_simulate, year_actual, s)
}
\arguments{
\item{years_to_simulate}{(int vector) years to simulate}

\item{year_actual}{(int) year actual of the model}

\item{s}{(int) loop number}
}
\value{
outputs of the years simulated by the antares runs
}
\description{
Wrapper of the run antares simulation : local and antares_web
}
