% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{run_simulation_antares}
\alias{run_simulation_antares}
\title{Run an antares simulation and modify automatically settings}
\usage{
run_simulation_antares(
  solver_path,
  simulation_path,
  simulation_name,
  antares_mode,
  sub_mode,
  mc_list,
  antares_run_save
)
}
\arguments{
\item{solver_path}{antares solver path}

\item{simulation_path}{path of the simulation run}

\item{simulation_name}{name of the simulation output antares}

\item{antares_mode}{default : economy}

\item{sub_mode}{antares optimisation mode fast or accurate}

\item{mc_list}{montecarlo playlist}

\item{antares_run_save}{option to make a run antares not allowed for deleting automatically}
}
\value{
the output name of the antares run
}
\description{
Run an antares simulation and modify automatically settings
}
