% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_web_utils.R
\name{run_simulation_antares_web}
\alias{run_simulation_antares_web}
\title{Run a simulation on antares web, specifiying the general options}
\usage{
run_simulation_antares_web(
  year,
  simulation_name,
  antares_mode,
  sub_mode,
  mc_list,
  verbose = TRUE
)
}
\arguments{
\item{year}{year of simulation}

\item{simulation_name}{name of simulation}

\item{antares_mode}{mode of simulation (economy)}

\item{sub_mode}{sub_mode (fast)}

\item{mc_list}{mc_list to simulate}

\item{verbose}{verbose}
}
\value{
job_id : id of the job launched on antares web
}
\description{
Run a simulation on antares web, specifiying the general options
}
