% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_investment_model.R
\name{select_best_capacity}
\alias{select_best_capacity}
\title{Compute the best candidates based on the NPV projected.}
\usage{
select_best_capacity(
  metrics,
  year_actual,
  candidates_investment,
  LOLD,
  area,
  revenues_total_net = NA
)
}
\arguments{
\item{metrics}{(dataframe) of metrics for each clusters}

\item{year_actual}{(int) year actual}

\item{candidates_investment}{(dataframe) candidates to investments}

\item{LOLD}{(scalar) Loss of Load duration}

\item{area}{(str) country code}

\item{revenues_total_net}{(dataframe) of net revenues for each clusters by year}
}
\value{
best cluster's name or string 'end'
}
\description{
Compute the best candidates based on the NPV projected.
}
