% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_investment_model.R
\name{select_best_capacity_base}
\alias{select_best_capacity_base}
\title{Base part of select_best_capacity function}
\usage{
select_best_capacity_base(
  best_metric,
  ordered_metric,
  year_actual,
  revenues_total_net = NA
)
}
\arguments{
\item{best_metric}{value of the best metric}

\item{ordered_metric}{table of metrics ordered}

\item{revenues_total_net}{dataframe of net revenues for each clusters by year}
}
\value{
best_capacity name
}
\description{
Base part of select_best_capacity function
}
