% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_investment_model.R
\name{select_worst_capacity_base}
\alias{select_worst_capacity_base}
\title{Base part of select_worst_capacity function}
\usage{
select_worst_capacity_base(
  worst_revenue,
  ordered_revenues,
  revenues_total_net,
  year_actual
)
}
\arguments{
\item{worst_revenue}{table of metrics ordered}

\item{ordered_revenues}{value of the best metric}

\item{revenues_total_net}{dataframe of total net revenues by year and cluster}
}
\value{
best_capacity name
}
\description{
Base part of select_worst_capacity function
}
