% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_utils.R
\name{set_simulation_antares}
\alias{set_simulation_antares}
\title{Set simulation Path wrapper for Antares}
\usage{
set_simulation_antares(study_id, input_mode, antares_output_name = NA)
}
\arguments{
\item{study_id}{(str) antares simulation id}

\item{input_mode}{(str) if TRUE, setsimulation in reading mode}

\item{antares_output_name}{(str) compulsory if output mode}
}
\value{
opts (list) antares simulation options
}
\description{
Set simulation Path wrapper for Antares
}
