% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_main_utils.R
\name{setup_antares_invest}
\alias{setup_antares_invest}
\title{Setup of directories and variables relative to the good progress of the model}
\usage{
setup_antares_invest()
}
\value{
list of essential variables
}
\description{
Setup of directories and variables relative to the good progress of the model
}
