% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_CRM_utils.r
\name{simulation_auction}
\alias{simulation_auction}
\title{Meet supply and demand curves and simulate the auction}
\usage{
simulation_auction(market_info_supply, load, area)
}
\arguments{
\item{market_info_supply}{(dataframe) dataframe from compute_bids()}

\item{load}{(float)}

\item{area}{(str) country code}
}
\value{
auction info list
}
\description{
Meet supply and demand curves and simulate the auction
}
