% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_CRM.R
\name{simulation_mecanism_capacity}
\alias{simulation_mecanism_capacity}
\title{Capacity market simulation mother function}
\usage{
simulation_mecanism_capacity(
  year_simulated,
  obligation,
  power_plants,
  capacity_certified,
  clusters_not_supported,
  zone,
  market_info_supply,
  spot = NA,
  LOLD = NA,
  revenues = NA
)
}
\arguments{
\item{year_simulated}{(str) year simulated}

\item{obligation}{(scalar) obligation computed by obligation()}

\item{power_plants}{(dataframe) with costs hypothesis by clusters}

\item{capacity_certified}{(scalar) certified capacity by cluster}

\item{clusters_not_supported}{(str vector) clusters names candidates to the revenues analysis and investments.}

\item{zone}{(str) country code}

\item{market_info_supply}{(dataframe) dataframe of market supply side information}
}
\value{
auction and the supply side information.
}
\description{
Capacity market simulation mother function
}
