% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_analysis_revenues.R
\name{system_service_revenues_cluster}
\alias{system_service_revenues_cluster}
\title{Service system revenue computation}
\usage{
system_service_revenues_cluster(c, power_plants, zone)
}
\arguments{
\item{c}{(str) cluster name}

\item{power_plants}{(dataframe) with costs hypothesis}

\item{zone}{(str) country code}
}
\value{
Revenues, annual granularity from service system
}
\description{
Service system revenue computation
}
