% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_investment_utils.R
\name{update_clusters_filters}
\alias{update_clusters_filters}
\title{Update the clusters interesting with the candidates to decommissioning already mentioned in antares studies}
\usage{
update_clusters_filters(dataset, area)
}
\arguments{
\item{dataset}{the dataset coming out from the build dataset function}

\item{area}{(str) country code}
}
\value{
updated clusters interesting for every simulated years
}
\description{
Update the clusters interesting with the candidates to decommissioning already mentioned in antares studies
}
