% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_antares_powerplants.R
\name{update_powerplants}
\alias{update_powerplants}
\title{Update the power_plants dataframe using the power_plants_moves dataframe}
\usage{
update_powerplants(
  power_plants,
  area,
  year_simulated,
  year_actual,
  study_id,
  s
)
}
\arguments{
\item{power_plants}{(dataframe) table with the power fleet}

\item{area}{(str) country code}

\item{year_simulated}{(int) year simulated}

\item{year_actual}{(int) current simulated year}

\item{s}{loop}

\item{power_plants_actual}{(dataframe) table with the power fleet actual}
}
\value{
power_plants updated with power_plants_moves
}
\description{
Update the power_plants dataframe using the power_plants_moves dataframe
}
